Frederic Bouvier: - This doesn't do what the author though it would do:   if (modifiers&KEYMOD_RELEASED == 0) ...   (see below) - MSVC and type fixes
mf:
fixing this uncovered an old bug: the "key pressed" branch was
actually never executed -- always only "released". And that is why
the <repeatable> property didn't work. And *that* was the reason
why the Spitfire's starter didn't work with freeglut, and Vivian
had to write a nasty workaround for it.