Tatsuhiro NISHIOKA:
Finally I found the cause of DList stack overflow and off-carrier
aircraft problem.

The cause of the first one is that aip.ssgTransform of AICarrier are
unexpectedly registered on every reset in AIBase::init().
The second one is caused by ssgEntry related code in AICarrier::init().
So I made a patch for doing these process only in the first init()
calls.